Fix conversion of swap interval from QGLFormat to QSurfaceFormat
When converting a QGLFormat with an unset/default swap interval
we would end up with an invalid swap interval in the resulting
QSurfaceFormat (since -1 is used as the "default" in QGLFormat).

One result of this would be that we would not vsync rendering
with QGLWidget on Wayland by default, since we implement our
own blocking swapBuffers() on Wayland.

Note: This is irrelevant for Qt 6, since QGLWidget no longer
exists.

[ChangeLog][QtOpenGL] Fixed an issue where the default swap
interval of QGLWidget's surface format would be set to an
invalid value.

Task-number: QTBUG-81504
Change-Id: If90708bbd7975d7ae1d27fafa20f837543dd81ee
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>