Add common internal macros to request jni objects
There are multiple macros used to get/check and return with error
in case the requested object is not valid. These macros are defined
in multiple places and duplicated. This patch defines them in one place
and then they can be reused.

This macro expects a "char m_qtTag[]" variable to be defined in
the scope where the macro is used. That variable is used as a tag
for the error message printed when an error occur.

Another consecutive patch use the new macros over qtbase code.

Change-Id: Ibb8558d1229cec6dad9ec9da6e2635ea54fd18d6
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit 01d30e2d5331d0d15516ce3bf7559fcc19b69723)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>